requirements.txt
pyaes==1.6.1
pyasn1==0.3.7
PySocks==1.6.7
rsa==3.4.2
Telethon==0.11.5
